refactor: split browser_cmd.py into focused modules (fixes #856)
Split the 1,378-line naturo/cli/browser_cmd.py into focused submodules
under naturo/cli/_browser/, mirroring the established app_cmd.py split
(#832/#847).

- The 'browser' Click group lives in naturo/cli/_browser/_group.py (a
  standalone module with no dependency on the command submodules), so the
  focused command modules attach to it without forming an import cycle.
- The canonical _get_page helper stays in browser_cmd.py; submodules call
  browser_cmd._get_page(...) so existing @patch('naturo.cli.browser_cmd.
  _get_page') test targets keep working.
- browser_cmd.py re-exports the group and every command callable, so
  'from naturo.cli.browser_cmd import ...' keeps working.

Modules: navigation, elements, frames, waits, network, visual, stealth,
lifecycle, captcha. Adds tests/test_browser_cmd_split.py to lock the full
35-command surface and backward-compatible re-exports.

ruff + mypy (py3.12) clean on the changed files.
